Move Exceptions to Exception.
