+++ src/Data/SqlDataProvider.php
@@ -5,7 +5,7 @@ namespace Yiisoft\Db\Data;
 use yii\data\BaseDataProvider;
 use Yiisoft\Db\Connection\Connection;
 use Yiisoft\Db\Querys\Query;
-use Yiisoft\Db\Exceptions\InvalidConfigException;
+use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Expressions\Expression;
 
 /**
