+++ src/Exception/Exception.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Exceptions;
+namespace Yiisoft\Db\Exception;
 
 /**
  * Exceptions represents an exception that is caused by some DB-related operations.
