+++ src/Exception/InvalidConfigException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Exceptions;
+namespace Yiisoft\Db\Exception;
 
 /**
  * Class InvalidConfigException.
