+++ src/Exception/NotSupportedException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Exceptions;
+namespace Yiisoft\Db\Exception;
 
 /**
  * NotSupportedException represents an exception caused by accessing features that are not supported.
