+++ src/Querys/Conditions/SimpleCondition.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Querys\Conditions;
 
-use Yiisoft\Db\Exceptions\InvalidArgumentException;
+use Yiisoft\Db\Exception\InvalidArgumentException;
 
 /**
  * Class SimpleCondition represents a simple condition like `"column" operator value`.
