+++ src/Querys/QueryBuilder.php
@@ -6,13 +6,13 @@ namespace Yiisoft\Db\Querys;
 
 use Generator;
 use Yiisoft\Db\Connection\Connection;
-use Yiisoft\Db\Exceptions\Exception;
-use Yiisoft\Db\Exceptions\InvalidConfigException;
+use Yiisoft\Db\Exception\Exception;
+use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Querys\Conditions\ConditionInterface;
 use Yiisoft\Db\Constraint\Constraint;
 use Yiisoft\Db\Constraint\ConstraintFinderInterface;
-use Yiisoft\Db\Exceptions\InvalidArgumentException;
-use Yiisoft\Db\Exceptions\NotSupportedException;
+use Yiisoft\Db\Exception\InvalidArgumentException;
+use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Expressions\Expression;
 use Yiisoft\Db\Expressions\ExpressionBuilder;
 use Yiisoft\Db\Expressions\ExpressionBuilderInterface;
