+++ src/Querys/QueryTrait.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Querys;
 
-use Yiisoft\Db\Exceptions\NotSupportedException;
+use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Expressions\ExpressionInterface;
 
 /**
